// app/types/types.d.ts
export type Role = "USER" | "ADMIN" | "OWNER" | "BROKER"; // | 'TENANT'
// ...
export type OwnerDetails = {
  name: string;
  email: string;
  number: string;
  ownerId: string;
  listerOwnerId: string;
};

// --------------------------------------------------------------------------------
// INTERESTED ROOMS
export type InterestedRooms = {
  id: string;
  lister: Lister;
  rooms: string[];
  createdAt: string;
};

// --------------------------------------------------------------------------------
// ROOM
export type RoomAmenities =
  | "AC"
  | "WIFI"
  | "POOL"
  | "LIFT"
  | "OVEN"
  | "SOFA"
  | "GARDEN"
  | "GEYSER"
  | "FRIDGE"
  | "BALCONY"
  | "TERRACE"
  | "CHIMNEY"
  | "CAR PARK"
  | "BIKE PARK"
  | "24/7 WATER"
  | "FIRE SAFETY"
  | "POWER BACKUP"
  | "WASHING MANCHINE";
export type RoomType = "FLAT" | "1BHK" | "2BHK" | "3BHK" | "4BHK";
export type FurnishingStatus = "FURNISHED" | "SEMIFURNISHED" | "UNFURNISHED";

export type Room = {
  hall: number;
  city: string;
  title: string;
  price: number;
  kitchen: number;
  bedroom: number;
  location: string;
  bathroom: number;
  listerName: string;
  roomType: RoomType;
  minCapacity: number;
  maxCapacity: number;
  description: string;
  primaryContact: string;
  secondaryContact: string;
  direction: string | null;
  amenities: RoomAmenities[];
  furnishingStatus: FurnishingStatus;
  // ENUM
  postedBy: Role;
  // RELATION
  listerId: string;
};

export type RoomWithMedia = Room & {
  photos: File[];
  videos: FileList | null;
};

export type RoomWithMediaUrl = Room & {
  photos: string[];
  videos: string | null;
};

export type ListedRoom = RoomWithMediaUrl & {
  id: string;
  ownerId: string;
  ratings: number;
  verified: boolean;
  available: boolean;
  // DATE
  createdAt: string;
  updatedAt: string;
};